Make unloadOtherTabs aware of tab groups, and provide an option for whether its scope should extend to the entire window or just to the currently active tab group, with the latter as default.
This patch was originally written for BarTab, and more detailed history
of its development can be found at:

https://github.com/inverseparadox/BarTab/

Unfortunately, in BarTab Lite X the "entire window" case is badly
buggy. It is therefore not recommended to change that option away from
its default, unless you want to work on fixing that bug.